canperf: Update the format of TX IDs provided to cangen
- latest cangen introduced a new generation mode for
even IDs, identified through the 'e' option. Internally,
cangen checks only the first letter of the option to
determine the mode used, thus for IDs such as 'e4'
(which is used for CAN2Eth flow), the cangen tool
would generate frames with random, even IDs, instead
of using the aforementioned ID. An work-around for this
is to prefix the ID with a '0'.

Issue: GVIP-1124
Signed-off-by: George Mocanu <george.mocanu@nxp.com>